Reorganized and renamed folders for easier use. Will now pull to master and create new development branch.
