--- CSS/CSS.md
@@ -1,2 +1,0 @@
-#CSS
-This folder will hold the CSS files for the whole project.
