--- JS/JS.md
@@ -1,2 +1,0 @@
-#JS
-This folder will hold the JavaScript files used for the project.
