--- Mockups/Mockups.md
@@ -1,2 +1,0 @@
-#Mockups
-This folder will hold all the files related to mocking up the project.
