+++ installer/INSTALLER.md
@@ -1,0 +1,6 @@
+#INSTALLER
+This folder contains the installer.
+The installer is responsible for the initial setup of the database and populating basic configuration data in it.
+It is a self contained application that should not rely on any part of the rest of the project.
+All configuration that may be required during installation should be done via the installer's web UI.
+No manual configuration should be required by the user during installation.
