+++ js/JS.md
@@ -1,0 +1,2 @@
+#JS
+This folder holds javascript helpers
