+++ layout/layout.md
@@ -1,0 +1,3 @@
+#LAYOUT
+This folder contains a mockup layout of the web site in purely html and css
+The files in this folder should act as a model for what the end result should look like and will be used to compile templates and themes at a later time.
