+++ lib/hookr.php
