+++ lib/settings.php
